--- ids.py
@@ -1,11 +1,0 @@
-# IDs for my server
-# TODO: Put real ids in
-
-SERVER_GUILD_ID = 0
-INTRODUCTION_CHANNEL_ID = 0
-WELCOME_CHANNEL_ID = 0
-INFORMATION_CHANNEL_ID = 0
-
-INTRO_QUESTIONS = {
-
-}
