--- yeet
@@ -1,32 +1,0 @@
-> functions@ build /home/hamza/Downloads/BitchBot/functions
-> tsc
-
-✔  functions: Finished running predeploy script.
-i  functions: ensuring necessary APIs are enabled...
-✔  functions: all necessary APIs are enabled
-i  functions: preparing functions directory for uploading...
-i  functions: packaged functions (33.53 KB) for uploading
-✔  functions: functions folder uploaded successfully
-i  functions: creating Node.js 8 function quotesChannels(us-central1)...
-i  functions: creating Node.js 8 function setQuotesChannel(us-central1)...
-i  functions: updating Node.js 8 function allEmojis(us-central1)...
-i  functions: updating Node.js 8 function saveUserInfo(us-central1)...
-i  functions: updating Node.js 8 function getUserInfo(us-central1)...
-i  functions: updating Node.js 8 function epicEmojis(us-central1)...
-i  functions: updating Node.js 8 function roles(us-central1)...
-i  functions: updating Node.js 8 function getRoleIdForAssignment(us-central1)...
-✔  functions[quotesChannels(us-central1)]: Successful create operation. 
-Function URL (quotesChannels): https://us-central1-bitchbot-discordbot.cloudfunctions.net/quotesChannels
-✔  functions[setQuotesChannel(us-central1)]: Successful create operation. 
-Function URL (setQuotesChannel): https://us-central1-bitchbot-discordbot.cloudfunctions.net/setQuotesChannel
-✔  functions[allEmojis(us-central1)]: Successful update operation. 
-✔  functions[saveUserInfo(us-central1)]: Successful update operation. 
-✔  functions[getRoleIdForAssignment(us-central1)]: Successful update operation. 
-✔  functions[getUserInfo(us-central1)]: Successful update operation. 
-✔  functions[epicEmojis(us-central1)]: Successful update operation. 
-✔  functions[roles(us-central1)]: Successful update operation. 
-
-✔  Deploy complete!
-
-Project Console: https://console.firebase.google.com/project/bitchbot-discordbot/overview
-[hamza@Hamza functions]$
